Reorganize examples by method rather than net arch
